third_party/blink/web_tests/fast/frames/sandboxed-iframe-attribute-parsing-05-expected.txt

CONSOLE ERROR: Blocked form submission to 'javascript:top.disallowedFormSubmitted();' because the form's frame is sandboxed and the 'allow-forms' permission is not set.
Mixed case

PASS
